[webkitpy] Introduce GLibPort base class https://bugs.webkit.org/show_bug.cgi?id=244773
Reviewed by Jonathan Bedard.

Code shared between the WPE/GTK webkitpy ports is now de-duplicated. Rejoice. And with this patch
WPE gained support for running tests in valgrind, but that wasn't tested in years I'm afraid, so it
might need a follow-up revamp.

* Tools/Scripts/webkitpy/port/glib.py: Added.
(GLibPort):
(GLibPort.__init__):
(GLibPort.default_timeout_ms):
(GLibPort._built_executables_path):
(GLibPort._built_libraries_path):
(GLibPort.setup_test_run):
(GLibPort.setup_environ_for_server):
(GLibPort._get_crash_log):
* Tools/Scripts/webkitpy/port/gtk.py:
(GtkPort):
(GtkPort._driver_class):
(GtkPort.driver_stop_timeout):
(GtkPort.setup_environ_for_server):
(GtkPort.check_sys_deps):
(GtkPort.__init__): Deleted.
(GtkPort._built_executables_path): Deleted.
(GtkPort._built_libraries_path): Deleted.
(GtkPort.default_timeout_ms): Deleted.
(GtkPort.setup_test_run): Deleted.
(GtkPort._get_crash_log): Deleted.
* Tools/Scripts/webkitpy/port/wpe.py:
(WPEPort):
(WPEPort.__init__):
(WPEPort.setup_environ_for_server):
(WPEPort.test_expectations_file_position):
(WPEPort.default_timeout_ms): Deleted.
(WPEPort._built_executables_path): Deleted.
(WPEPort._built_libraries_path): Deleted.
(WPEPort._get_crash_log): Deleted.

Canonical link: https://commits.webkit.org/254212@main
