+++ Tools/Scripts/webkitpy/port/gtk.py
@@ -29,7 +29,6 @@
 # OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 import os
-import uuid
 import logging
 import shlex
 
@@ -37,43 +36,19 @@ import webkitpy
 from webkitpy.common.system import path
 from webkitpy.common.memoized import memoized
 from webkitpy.layout_tests.models.test_configuration import TestConfiguration
-from webkitpy.port.base import Port
+from webkitpy.port.glib import GLibPort
 from webkitpy.port.xvfbdriver import XvfbDriver
 from webkitpy.port.westondriver import WestonDriver
 from webkitpy.port.xorgdriver import XorgDriver
 from webkitpy.port.waylanddriver import WaylandDriver
-from webkitpy.port.linux_get_crash_log import GDBCrashLogGenerator
-from webkitpy.port.leakdetector_valgrind import LeakDetectorValgrind
 
 from webkitcorepy import decorators
 
 _log = logging.getLogger(__name__)
 
 
-class GtkPort(Port):
+class GtkPort(GLibPort):
     port_name = "gtk"
-
-    def __init__(self, *args, **kwargs):
-        super(GtkPort, self).__init__(*args, **kwargs)
-        self._display_server = self.get_option("display_server")
-
-        if self.get_option("leaks"):
-            self._leakdetector = LeakDetectorValgrind(self._executive, self._filesystem, self.results_directory())
-            if not self.get_option("wrapper"):
-                raise ValueError('use --wrapper=\"valgrind\" for memory leak detection on GTK')
-
-        if self._should_use_jhbuild():
-            self._jhbuild_wrapper = [self.path_from_webkit_base('Tools', 'jhbuild', 'jhbuild-wrapper'), '--gtk', 'run']
-            if self.get_option('wrapper'):
-                self.set_option('wrapper', ' '.join(self._jhbuild_wrapper) + ' ' + self.get_option('wrapper'))
-            else:
-                self.set_option_default('wrapper', ' '.join(self._jhbuild_wrapper))
-
-    def _built_executables_path(self, *path):
-        return self._build_path(*(('bin',) + path))
-
-    def _built_libraries_path(self, *path):
-        return self._build_path(*(('lib',) + path))
 
     def _port_flag_for_scripts(self):
         return "--gtk"
@@ -88,50 +63,15 @@ class GtkPort(Port):
             return XorgDriver
         return XvfbDriver
 
-    def default_timeout_ms(self):
-        default_timeout = 15000
-        # Starting an application under Valgrind takes a lot longer than normal
-        # so increase the timeout (empirically 10x is enough to avoid timeouts).
-        multiplier = 10 if self.get_option("leaks") else 1
-        # Debug builds are slower (no compiler optimizations are used).
-        if self.get_option('configuration') == 'Debug':
-            multiplier *= 2
-        return multiplier * default_timeout
-
     def driver_stop_timeout(self):
         if self.get_option("leaks"):
             # Wait the default timeout time before killing the process in driver.stop().
             return self.default_timeout_ms()
         return super(GtkPort, self).driver_stop_timeout()
 
-    def setup_test_run(self, device_type=None):
-        super(GtkPort, self).setup_test_run(device_type)
-
-        if self.get_option("leaks"):
-            self._leakdetector.clean_leaks_files_from_results_directory()
-
     def setup_environ_for_server(self, server_name=None):
         environment = super(GtkPort, self).setup_environ_for_server(server_name)
-        environment['G_DEBUG'] = 'fatal-criticals'
-        environment['GSETTINGS_BACKEND'] = 'memory'
         environment['LIBOVERLAY_SCROLLBAR'] = '0'
-        environment['TEST_RUNNER_INJECTED_BUNDLE_FILENAME'] = self._build_path('lib', 'libTestRunnerInjectedBundle.so')
-        environment['TEST_RUNNER_TEST_PLUGIN_PATH'] = self._build_path('lib', 'plugins')
-        self._copy_value_from_environ_if_set(environment, 'WEBKIT_OUTPUTDIR')
-        self._copy_value_from_environ_if_set(environment, 'WEBKIT_JHBUILD')
-        self._copy_value_from_environ_if_set(environment, 'WEBKIT_TOP_LEVEL')
-        self._copy_value_from_environ_if_set(environment, 'WEBKIT_DEBUG')
-        self._copy_value_from_environ_if_set(environment, 'WEBKIT_GST_USE_PLAYBIN3')
-        self._copy_value_from_environ_if_set(environment, 'AT_SPI_BUS_ADDRESS')
-        for gst_variable in ('DEBUG', 'DEBUG_DUMP_DOT_DIR', 'DEBUG_FILE', 'DEBUG_NO_COLOR',
-                             'PLUGIN_SCANNER', 'PLUGIN_PATH', 'PLUGIN_SYSTEM_PATH', 'REGISTRY',
-                             'PLUGIN_PATH_1_0'):
-            self._copy_value_from_environ_if_set(environment, 'GST_%s' % gst_variable)
-
-        gst_feature_rank_override = os.environ.get('GST_PLUGIN_FEATURE_RANK')
-        environment['GST_PLUGIN_FEATURE_RANK'] = 'fakeaudiosink:max'
-        if gst_feature_rank_override:
-            environment['GST_PLUGIN_FEATURE_RANK'] += ',%s' % gst_feature_rank_override
 
         # Configure the software libgl renderer if jhbuild ready and we test inside a virtualized window system
         if self._driver_class() in [XvfbDriver, WestonDriver] and (self._should_use_jhbuild() or self._is_flatpak()):
@@ -151,31 +91,6 @@ class GtkPort(Port):
                 environment['LIBGL_DRIVERS_PATH'] = dri_libgl_path
             else:
                 _log.warning("Can't find Gallium llvmpipe driver. Try to run update-webkitgtk-libs or update-webkit-flatpak")
-        if self.get_option("leaks"):
-            # Turn off GLib memory optimisations https://wiki.gnome.org/Valgrind.
-            environment['G_SLICE'] = 'always-malloc'
-            environment['G_DEBUG'] += ',gc-friendly'
-            # Turn off bmalloc when running under Valgrind, see https://bugs.webkit.org/show_bug.cgi?id=177745
-            environment['Malloc'] = '1'
-            xmlfilename = "".join(("drt-%p-", uuid.uuid1().hex, "-leaks.xml"))
-            xmlfile = os.path.join(self.results_directory(), xmlfilename)
-            suppressionsfile = self.path_from_webkit_base('Tools', 'Scripts', 'valgrind', 'suppressions.txt')
-            environment['VALGRIND_OPTS'] = \
-                "--tool=memcheck " \
-                "--num-callers=40 " \
-                "--demangle=no " \
-                "--trace-children=no " \
-                "--smc-check=all-non-file " \
-                "--leak-check=yes " \
-                "--leak-resolution=high " \
-                "--show-possibly-lost=no " \
-                "--show-reachable=no " \
-                "--leak-check=full " \
-                "--undef-value-errors=no " \
-                "--gen-suppressions=all " \
-                "--xml=yes " \
-                "--xml-file=%s " \
-                "--suppressions=%s" % (xmlfile, suppressionsfile)
         return environment
 
     def _generate_all_test_configurations(self):
@@ -243,10 +158,6 @@ class GtkPort(Port):
 
     def check_sys_deps(self):
         return super(GtkPort, self).check_sys_deps() and self._driver_class().check_driver(self)
-
-    def _get_crash_log(self, name, pid, stdout, stderr, newer_than, target_host=None):
-        return GDBCrashLogGenerator(self._executive, name, pid, newer_than,
-                                    self._filesystem, self._path_to_driver, self.port_name, self.get_option('configuration')).generate_crash_log(stdout, stderr)
 
     def test_expectations_file_position(self):
         # GTK port baseline search path is gtk -> glib -> wk2 -> generic (as gtk-wk2 and gtk baselines are merged), so port test expectations file is at third to last position.
