+++ Tools/Scripts/webkitpy/port/wpe.py
@@ -33,42 +33,21 @@ import shlex
 from webkitpy.common.system import path
 from webkitpy.common.memoized import memoized
 from webkitpy.layout_tests.models.test_configuration import TestConfiguration
-from webkitpy.port.base import Port
+from webkitpy.port.glib import GLibPort
 from webkitpy.port.headlessdriver import HeadlessDriver
-from webkitpy.port.linux_get_crash_log import GDBCrashLogGenerator
 
 from webkitcorepy import decorators
 
 
-class WPEPort(Port):
+class WPEPort(GLibPort):
     port_name = "wpe"
 
     def __init__(self, *args, **kwargs):
         super(WPEPort, self).__init__(*args, **kwargs)
 
-        self._display_server = self.get_option("display_server")
         if self._display_server == 'xvfb':
             # While not supported by WPE, xvfb is used as the default value in the main scripts
             self._display_server = 'headless'
-        if self._should_use_jhbuild():
-            self._jhbuild_wrapper = [self.path_from_webkit_base('Tools', 'jhbuild', 'jhbuild-wrapper'), '--wpe', 'run']
-            self.set_option_default('wrapper', ' '.join(self._jhbuild_wrapper))
-
-    def default_timeout_ms(self):
-        default_timeout = 15000
-        # Starting an application under Valgrind takes a lot longer than normal
-        # so increase the timeout (empirically 10x is enough to avoid timeouts).
-        multiplier = 10 if self.get_option("leaks") else 1
-        # Debug builds are slower (no compiler optimizations are used).
-        if self.get_option('configuration') == 'Debug':
-            multiplier *= 2
-        return multiplier * default_timeout
-
-    def _built_executables_path(self, *path):
-        return self._build_path(*(('bin',) + path))
-
-    def _built_libraries_path(self, *path):
-        return self._build_path(*(('lib',) + path))
 
     def _port_flag_for_scripts(self):
         return "--wpe"
@@ -79,32 +58,8 @@ class WPEPort(Port):
 
     def setup_environ_for_server(self, server_name=None):
         environment = super(WPEPort, self).setup_environ_for_server(server_name)
-        environment['G_DEBUG'] = 'fatal-criticals'
-        environment['GSETTINGS_BACKEND'] = 'memory'
-        environment['TEST_RUNNER_INJECTED_BUNDLE_FILENAME'] = self._build_path('lib', 'libTestRunnerInjectedBundle.so')
-        environment['TEST_RUNNER_TEST_PLUGIN_PATH'] = self._build_path('lib', 'plugins')
-        environment['WEBKIT_EXEC_PATH'] = self._build_path('bin')
-        self._copy_value_from_environ_if_set(environment, 'WEBKIT_OUTPUTDIR')
-        self._copy_value_from_environ_if_set(environment, 'WEBKIT_JHBUILD')
-        self._copy_value_from_environ_if_set(environment, 'WEBKIT_TOP_LEVEL')
-        self._copy_value_from_environ_if_set(environment, 'WEBKIT_DEBUG')
-        self._copy_value_from_environ_if_set(environment, 'AT_SPI_BUS_ADDRESS')
-        self._copy_value_from_environ_if_set(environment, 'LIBGL_ALWAYS_SOFTWARE')
-        self._copy_value_from_environ_if_set(environment, 'PULSE_SERVER')
-        self._copy_value_from_environ_if_set(environment, 'PULSE_CLIENTCONFIG')
         self._copy_value_from_environ_if_set(environment, 'XR_RUNTIME_JSON')
-        self._copy_value_from_environ_if_set(environment, 'WEBKIT_GST_USE_PLAYBIN3')
         self._copy_value_from_environ_if_set(environment, 'BREAKPAD_MINIDUMP_DIR')
-        for gst_variable in ('DEBUG', 'DEBUG_DUMP_DOT_DIR', 'DEBUG_FILE', 'DEBUG_NO_COLOR',
-                             'PLUGIN_SCANNER', 'PLUGIN_PATH', 'PLUGIN_SYSTEM_PATH', 'REGISTRY',
-                             'PLUGIN_PATH_1_0'):
-            self._copy_value_from_environ_if_set(environment, 'GST_%s' % gst_variable)
-
-        gst_feature_rank_override = os.environ.get('GST_PLUGIN_FEATURE_RANK')
-        environment['GST_PLUGIN_FEATURE_RANK'] = 'fakeaudiosink:max'
-        if gst_feature_rank_override:
-            environment['GST_PLUGIN_FEATURE_RANK'] += ',%s' % gst_feature_rank_override
-
         return environment
 
     def show_results_html_file(self, results_filename):
@@ -141,10 +96,6 @@ class WPEPort(Port):
     def test_expectations_file_position(self):
         # WPE port baseline search path is wpe -> glib -> wk2 -> generic, so port test expectations file is at third to last position.
         return 3
-
-    def _get_crash_log(self, name, pid, stdout, stderr, newer_than, target_host=None):
-        return GDBCrashLogGenerator(self._executive, name, pid, newer_than,
-                                    self._filesystem, self._path_to_driver, self.port_name, self.get_option('configuration')).generate_crash_log(stdout, stderr)
 
     def configuration_for_upload(self, host=None):
         configuration = super(WPEPort, self).configuration_for_upload(host=host)
